refactored to be more versitile and efficient
